Заменил в TeamId ushort на byte.

+++ Assets/Scripts/ScriptableObjects/CollidableObjects/EntityCreatorObject.cs
@@ -13,7 +13,7 @@ public abstract class EntityCreatorObject : ScriptableObject
         return entity;
     }
 
-    public GameEntity CreateEntity(GameContext context, ushort teamId)
+    public GameEntity CreateEntity(GameContext context, byte teamId)
     {
         var entity = CreateEntity(context);
         foreach (var child in entity.GetAllChildrenGameEntities(context))
@@ -23,7 +23,7 @@ public abstract class EntityCreatorObject : ScriptableObject
         return entity;
     }
 
-    public GameEntity CreateEntity(GameContext context, Vector2 position, float angle, ushort teamId)
+    public GameEntity CreateEntity(GameContext context, Vector2 position, float angle, byte teamId)
     {
         var entity = CreateEntity(context, position, angle);
         foreach (var child in entity.GetAllChildrenGameEntities(context))
